idl: No need to show ErrorInfo? return type
The spec states: “Values produced by I/O (e.g. a request) are prefixed
with @=> io@. In some platforms (JS) those values are passed as
arguments to a callback argument, instead of being returned. I/O always
can fail, but how do they fail is undefined in the spec, so it's also
undefined here”

As such, it is assumed that the developer implementing this IDL will
implicitly figure out how to deal with io failures such as using
callbacks with an err argument in JS, or a CompletionListener in Java