spec: Remove ambiguity about the value for @resumed@ emitted by auto-reenter
This is a tricky one because if we set resumed to false, then it implies that the user has lost continuity on the channel and should get history for example if they want to get the missing messages.  However, this is not the case as this is a presence re-enter failure. 

Instead resumed should be true because the channel is still providing continuity, and instead the local presence state has fallen out of sync, and thus it is up to the developer to handle this edge case. We should add a note in the user documentation about this condition and the fact that an error in the range 91000 -> 91099 will be emitted.