spec: Make ProtocolMessage private.
ProtocolMessage, as a structured type, is in the spec, yet it isn't
referred to by any other public part of the interface. It's only
mentioned as a type used by internal operations.

We could just remove it (as it's already defined in the Realtime
Protocol Definition, where it really belongs), but I think there's value
in keeping it here in a higher-level form to keep library codebases
uniform.