RTN11c: Explicit connect should keep trying to recover connection state.
"Aborts the retry process" suggests that we should just abandon the
whole connection state recovery logic. In implementation terms, this
may be translated to breaking an actual loop with the associated
connection state TTL ([this happened recently in Go](https://github.com/ably/ably-go/pull/297)).

This doesn't make much sense, and it's really not useful. What should
happen instead is that the reconnection loop should skip the wait, and
attempt the next try right away.