Add spec points for RealtimeClient#close/connect
This was already in the IDL, and the behaviour described there (proxy to
method on underlying Connection) matches the behaviour in a few of our
existing SDKs that I looked at:

- ably-js@0e4e590
- ably-java@faa89af
- ably-ruby@65e0a24

Strangely, ably-cocoa@05f9eb8 does the opposite - all of the
connection-related behaviour is implemented in ARTRealtime, and then
ARTConnection just proxies method calls through to the corresponding
ARTRealtime instance. But I suppose that from a public API point of view
the final result is the same.