Add feature spec point for RealtimeClient#clientId
tl;dr: I’m feeling a bit confused and 🤷 about this one.

This already existed in the IDL, and is implemented in a few of our
client SDKs.

I had to guess a bit as to the intended behaviour here. The only
existing clue in the feature spec is the IDL’s “proxy for RSA7”, which
isn’t that helpful since RSA7 doesn’t describe a method or attribute –
it describes a behaviour (that of determining which clientId to use when
requesting a token). So perhaps you could argue that it means to say
that #clientId should return this value, which is essentially the same
as the clientId provided in the ClientOptions.

But then you also have RSA12 and RSA17, which describe the behaviour of
the Auth#clientId attribute, which is slightly different – essentially,
it initially returns the value provided in the ClientOptions, but it can
become null if, subsequently, "@ConnectionDetails#clientId@ is @null@
following a connection to Ably". (To be honest, I don’t really
understand what are the circumstances under which the ClientOptions
would specify a non-null clientId, and then a subsequent connection to
Ably would return a null clientId in the ConnectionDetails.)

It probably doesn’t help that I still don’t fully understand our auth
flow, full stop, so this might need some external input.

So, to try and decide which behaviour was the intended one, I looked at
some of our client SDKs that implement this attribute:

- ably-cocoa@05f9eb8: returns the ClientOptions’s clientId
- ably-ruby@65e0a24: proxy to Auth#clientId
- ably-dotnet@9a3c31e: proxy to Auth#clientId

So, on the very slight balance of numbers, and since I think it makes
sense to go for the more dynamic of the two options, I’m opting for the
proxy to Auth#clientId. I don’t really understand _why_ we’d want to
have this attribute that just proxies an Auth one, given that the Auth
object is already exposed publicly on RealtimeClient.