Add ClientInformation class and .agents property
This gives us access to all of the information that the library uses to form
its Agent identifier. For example, it allows a user of the library to
programmatically find out which version of the Ably client library is being
used at runtime.

The use case that triggered this addition is that our Asset Tracking SDKs would
like to be able to attach Ably client library version information to data that
they archive [1].

I’ve opted to add a new class since we didn’t have an obvious place to put this
sort of information. There are a couple of non-spec examples that exist in our
SDKs - namely ARTDefault in ably-cocoa and Defaults in ably-java. Quintin had
suggested [2] we might want to formalise these existing examples, but I think
that their names aren’t particularly descriptive, and I don’t think that two
examples is enough to consider it a pattern that we should formalise.

I’ve made .agents a property since I’m thinking of a property as something that
could have custom getter logic. Languages that don’t have this feature might
need to implement it as a static method instead.

[1] https://github.com/ably/ably-asset-tracking-common/pull/66#issuecomment-1309993388
[2] https://github.com/ably/specification/issues/111#issuecomment-1311433432