Re-add mandatory version param for RestClient#request
This reverts commit 285e375 (but using the latest guidance for how to document
a replaced spec point).

We introduced this functionality in 1197e58 but subsequently removed it in
285e375 after deciding that we did not want to make breaking API changes in
specification version 2. Now that we’re starting work on version 3, we can
reintroduce it.

Same comment as 69e1973 re needing to update "replaced in" spec version before
merging integration branch into `main` applies here too.