Remove Connection#recoveryKey
This partially reverts commit 63affff.

We removed this property in cf35f3d but subsequently re-introduced it in
63affff after deciding that we did not want to make breaking API changes in
specification version 2. Now that we’re starting work on version 3, we can
remove it again.

In line with the guidance of CONTRIBUTING.md, which says that a specification
change that necessitates a major specification version increase should be
accompanied by said increase, here we bump the major version to 3.

Note that since we’re currently working on an integration branch for
specification version 3 and development is still ongoing on version 2, the
specification version number mentioned in RTN16m* may need to be updated before
merging the integration branch into `main`.