Explain what to do with RestClient#request’s version param
There’s still a question of how this method is meant to behave if the user also
specifies an `X-Ably-Header` header in the `headers` parameter. But it’s a
question that equally applied before adding the `version` parameter, so I’ve
captured it in a separate issue #144.

Same comment as 69e1973 re needing to update "replaced in" spec version before
merging integration branch into `main` applies here too.