Preserve pre-untyped stats spec points
This attempts to minimise the changes of a731d12, to avoid unnecessary spec
point churn and make the 2.1.0...3.0.0 diff more readable.

I’m assuming the introduction of TS15 was done to avoid having a load of
deletion notices for the TS12d-o subpoints, which is a valid aim, but I think
overall it’s better this way.