Update field names for mutable messages
Per discussion and subsequent call
https://ably-real-time.slack.com/archives/C02NY1VT3LY/p1732031493162379

Didn't bother deprecating and assigning new spec names since it's only
been implemented by us in one sdk so far, we can just retcon

One thing we didn't discuss is whether `version` or `serial` should be
the 'primary' field and which should be the derived one (copied over
from the other by the sdk in the case of CREATEs). Decided to make
`version` the primary field, firstly because `timestamp` is already a
primary field, and that now corresponds with the version not the serial,
and secondly because that'll mean we might be able to skip sending the
version in updates/deletes.