Add API for tracking usage of wrapper SDKs
This specifies the API agreed in ADR-117 [1], to allow us to track usage of a
"modern wrapper SDK" (the term used in that ADR to refer to a wrapper SDK which
accepts an already-instantiated core SDK client). We will use it to track usage
of the Chat SDK.

The functionality described here has already been partially implemented in
ably-cocoa [2] and is being used by ably-chat-swift [3].

I found it quite hard to write a spec for this, and the language might still be
both too vague and too convoluted. I'm putting this up for review as a first
attempt in order not to end up getting stuck trying to write something perfect.

[1] https://ably.atlassian.net/wiki/spaces/ENG/pages/3413016589/ADR-117+Tracking+wrapper+SDK+usage+continued
[2] https://github.com/ably/ably-cocoa/pull/2014
[3] https://github.com/ably/ably-chat-swift/pull/211