Do not fail ACK-pending messages for a failed resume
It contradicts RTN19a2 (if per RTN7c we're dropping ACK-pending messages
when the connection state is lost, then there are never going to be any
messages to assign a new msgSerial to in the second part of RTN19a2).

I wasn't sure what the intended behaviour was — I thought that perhaps
we were intentionally dropping ACK-pending messages after a failed
resume so as not to retry publishes if we couldn't guarantee they'd be
idempotent, but Simon confirmed in #289 that this is not the intention:

> This clause should be deleted. It's partly redundant (to SUSPENDED), and
> where it isn't redundant (the case you mentioned of a failed resume)
> it's wrong.
>
> We do retry pending messages after a failed resume (e.g. a different
> region) and that is indeed a case where you don't get idempotency unless
> you're using client-supplied message ids. (This is not well documented
> right now, very old docs ticket to improve that is ably/docs#807 /
> https://ably.atlassian.net/browse/EDU-1762 )

(There's a subsequent discussion in #289 about the various things we
might be trying to express when we say "idempotent", which I still need
to return to.)

I did a quick search for RTN7c in ably-js, ably-cocoa, and ably-java,
and it appears that, indeed, failure of ACK-pending messages is driven
purely by connection state.

Resolves #289.