Make `endpoint` special-case `localhost` and IP addresses
I believe it's quite common (e.g. for Realtime devs) to want to point to
a local Realtime. Or, for example in the SDK team, to want to point to a
local reverse proxy for interception.

Turns out that it's already special-cased in ably-go [1], so let's add
to spec.

Since ably-go is the only place that the `endpoint` functionality has
been implemented so far, I'm just changing the existing spec point and
will update ably-go to reflect.

[1] https://github.com/ably/ably-go/blob/17f6575773b28a146ac04bfaf3b8de44ce1ff162/ably/options.go#L506-L510