Clarify what state to transition to on #detach timeout
Since this behaviour relates to what to do when you've sent a DETACH but
haven't yet received a DETACHED, I want to make sure I understand it
before moving on to considering the same thing in the context of some
not-currently-handled connection state changes.

The current language of "previous state" is a bit vague to me — I think
that you can interpret it as meaning either the start-of-RTL5d ATTACHED
or the triggered-by-RTL5d DETACHING.

ably-js [1] and ably-cocoa [2] transition back to ATTACHED so that's the
behaviour that I've specified here, but I don't fully understand the
logic — it seems a bit misleading that you'd describe a channel as
ATTACHED when in fact the detach request may have been honoured (in
which case Realtime is not going to send you any further messages on
that channel).

(TODO: Would be good to get some insight on this one from someone
involved in this decision, so that I can add an explanatory comment. I
guess the thinking might be that the only way you're not going to
receive the DETACHED is because of a connection issue, in which case
you're eventually going to try reattaching and your ATTACHED state will
not be a lie any more. And if you receive the DETACHED late, you'll
reattach per RTL13)

[1] https://github.com/ably/ably-js/blob/b5fbb90c805d560813eec0637f140c39106dd105/src/common/lib/client/realtimechannel.ts#L842-L846
[2] https://github.com/ably/ably-cocoa/blob/ebe0e752e8d35ac9cceb0b4f2508675a4d519e15/Source/ARTRealtimeChannel.m#L1025