Add side effects of more connection states for DETACHING channels
We add handling for the following connection states:

FAILED: We're never going to receive the DETACHED that we're waiting
for, but we don't want to remain in DETACHING forever. So, let's put the
channel into FAILED and fail the #detach request.

CLOSED: We're never going to receive the DETACHED that we're waiting
for, but we don't want to remain in DETACHING forever. I don't think we
should consider this as a failure of the #detach request, because by
becoming CLOSED we've decided that we consider the server-side
connection state to have been torn down, and so it seems like we should
also consider the server-side attachment state to have also been torn
down. So, succeed the #detach request.

SUSPENDED: By becoming SUSPENDED, we've decided that there is no longer
any server-side connection state, and so I think we can conclude there
is also no longer any server-side attachment state and thus consider the
request to tear it down (that is, the #detach request) to have
succeeded. It's also consistent with what we do in RTL5j when #detach is
called on a SUSPENDED channel.

Now that there are more ways in which a #detach request can resolve, we
bring #detach in line with #attach and make its callback be driven by
the channel state.

Note: we noticed the need for these spec points because we noticed ([1],
[2]) that, in the absence of guidance from the spec, in ably-js a
DETACHING channel actually ends up becoming SUSPENDED when the
connection becomes SUSPENDED, which means it ends up reattaching when
the connection becomes CONNECTED again. This is not what we want — a
DETACHING channel should never autonomously reattach.

[1] https://ably-real-time.slack.com/archives/C8SPU4589/p1732273028523529?thread_ts=1732191641.458019&cid=C8SPU4589
[2] https://github.com/ably/ably-js/blob/18a255948c38d1e60715c8f5d6173369b57cb8d6/src/common/lib/client/baserealtime.ts#L178