Spec: specify that the delta bases are only MESSAGE_CREATEs
most obviously, not META occupancy etc (those obviously have no payloads and
would otherwise break the delta stream)

also exempt summaries, updates, and deletes, which (a) are (at least for now)
not order-preserved relative to the create stream from a single publisher,
and (b) are probably modifying (or deleting) old messages. Honestly, mutable
messages and deltas are an unlikely combination -- people using deltas are
probably just using creates as updates, sending a large constantly-changing
data object as a stream of new messages, so probably aren't going to want to
mutate old payloads. But if someone does have a usecase for it, this seems to
me like the most sensible approach.