Add plugin spec for Objects
Also renames feature spec for `Objects` to `RealtimeObjects`.
In the future we may add the REST API support for Objects, and it is
only reasonable to name that `RestObjects`. To stay consistent with the
namings for other features (e.g. RestChannels - RealtimeChannels) we
should rename the spec for `Objects` to `RealtimeObjects`