Clarify Chat URL-encoding requirements
"URL-encoded" is a bit vague because URLs allow different characters
depending on which component of the URL you're talking about; this led
to confusion when implementing this requirement in Swift.

At Andy's suggestion, I've also updated the spec to state that all path
variables must be escaped — "saves us having to write a new point every
time theres a URL param".