objects: Specify SYNCING and SYNCED events
Based on JS at e280bff and Swift at 1c901a8. (Got Claude to write this
spec based on the code but it needed a fair bit of editing afterwards.)

The change to spec point RTO8a doesn't match the Swift behaviour but it
does match Kotlin and JS; will address for Swift in [1].

Note that we've used the new SubscriptionStatus naming (introduced in
path-based API) instead of OnObjectsEventResponse; done at Andrii's
request in code review.

[1] https://ably.atlassian.net/browse/AIT-254